+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## Unreleased
+## 78.0 (2024-08-13)
 
 Compare: <https://git.torchbox.com/buckinghamshire-council/bc/compare/77.0...78.0>